backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.19
psycopg2-binary==2.9.9
python-jose==3.1.0
passlib[bcrypt]==1.7.4
pydantic==2.7.0
pydantic-settings==2.1.0
alembic==1.11.2
pytest==7.4.1
